feat: created one config
