+++ package.json
@@ -22,23 +22,11 @@
     "configuration": {
       "title": "Bothrs Sidekick",
       "properties": {
-        "bothrs.productTeam": {
-          "type": "string",
-          "default": "ambient",
-          "enum": [
-            "ambient",
-            "commerce",
-            "wap"
-          ],
-          "description": "Specify which Bothrs team's standard to use. Available options are \"ambient\", \"commerce\", \"wap\"."
-        },
         "bothrs.framework": {
-          "type": [
-            "string",
-            "null"
-          ],
-          "default": null,
+          "type": "string",
+          "default": "automatic",
           "enum": [
+            "automatic",
             "react",
             "react-native"
           ],
@@ -46,7 +34,7 @@
         },
         "bothrs.shouldOpenCreatedComponentFile": {
           "type": "boolean",
-          "default": false,
+          "default": true,
           "description": "Specify if Bothrs Sidekick should open the created component file."
         }
       }
