--- src/modules/createComponent/index.ts
@@ -1,34 +1,0 @@
-import { Framework, ProductTeam } from '../../types/configuration'
-
-import { createAmbientComponent } from './createAmbientComponent'
-import { createFunctionalComponent } from './createFunctionalComponent'
-
-import type { Uri } from 'vscode'
-
-export const createComponent = async (
-  commandPath: string,
-  componentName: string,
-  framework: Framework,
-  productTeam: ProductTeam
-): Promise<Uri | void> => {
-  const sanitizedComponentName =
-    componentName.charAt(0).toUpperCase() + componentName.slice(1)
-
-  if (productTeam === ProductTeam.Ambient) {
-    return await createAmbientComponent(
-      commandPath,
-      sanitizedComponentName,
-      framework
-    )
-  } else if (productTeam === ProductTeam.Commerce) {
-    // To be replaced with commerce standard
-    await createAmbientComponent(commandPath, sanitizedComponentName, framework)
-    return
-  }
-
-  return await createFunctionalComponent(
-    commandPath,
-    sanitizedComponentName,
-    framework
-  )
-}
