+++ src/types/configuration.ts
@@ -1,16 +1,10 @@
 export enum Framework {
+  Automatic = 'automatic',
   React = 'react',
   ReactNative = 'react-native',
 }
 
-export enum ProductTeam {
-  Ambient = 'ambient',
-  WAP = 'wap',
-  Commerce = 'commerce',
-}
-
 export interface ExtensionConfig {
   framework: Framework
-  productTeam: ProductTeam
   shouldOpenCreatedComponentFile: boolean
 }
